FIX: better highliting of schedule on tracks
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@5164 8aca7d54-2c30-db11-9de9-000461428c89

+++ boden/boden.cc
@@ -31,7 +31,7 @@ boden_t::boden_t(karte_t *welt, loadsave_t *file, koord pos ) : grund_t( welt, k
 			file->rdwr_long( age );
 			// check, if we still have this tree ... (if there are not trees, the first index is NULL!)
 			if(  baum_t::get_anzahl_besch()>id  &&  (*(baum_t::get_all_besch()))[id]!=NULL  ) {
-				baum_t *tree = new baum_t( welt, get_pos(), (uint16)id, age, slope );
+				baum_t *tree = new baum_t( welt, get_pos(), (uint8)id, age, slope );
 				dinge.add( tree );
 			}
 			else {
